Wait for all database tasks
This change updates the cross depends for the keystone-db task to ensure
all database tasks are run prior to running keystone db creation.
Previously we were only waiting for the primary controller database
tasks but if this plugin is used in conjunction with the
detached-database plugin it can cause failures.

Change-Id: I4e1068306c7b8d568e41acdcf0692b6c75e5b04f
Closes-Bug: #1595219